fix(stagewise): polish settings sidebar navigation
