+++ apps/browser/src/ui/screens/main/_components/sidebar-panel-config.ts
@@ -1,0 +1,7 @@
+export const SIDEBAR_PANEL_ID = 'new-sidebar-panel';
+export const SIDEBAR_PANEL_ORDER = 0;
+export const DEFAULT_EXPANDED_SIDEBAR_SIZE = 12;
+export const SIDEBAR_PANEL_MIN_SIZE = 6;
+export const SIDEBAR_PANEL_MAX_SIZE = 50;
+export const SIDEBAR_PANEL_CLASS_NAME =
+  '@container group overflow-visible! relative flex h-full min-w-44 max-w-2xl flex-col items-stretch';
